+++ tests/testthat/test-verb-slice.R
@@ -5,6 +5,15 @@ test_that("slice, head and tail aren't available", {
   expect_snapshot(error = TRUE, lf |> slice_tail())
 })
 
+test_that("min, max, and sample generate useful sql", {
+  lf <- lazy_frame(x = 1)
+  expect_snapshot({
+    lf |> slice_min(x, n = 1)
+    lf |> slice_max(x, prop = 0.5)
+    lf |> slice_sample(x, n = 1)
+  })
+})
+
 test_that("slice_min handles arguments", {
   db <- local_memdb_frame(x = c(1, 1, 2), id = c(1, 2, 3))
 
@@ -31,6 +40,12 @@ test_that("slice_max orders in opposite order", {
   expect_snapshot(error = TRUE, db |> slice_max(id, n = 1, na_rm = FALSE))
 })
 
+test_that("slice_min and slice_max strip missing values", {
+  db <- memdb_frame(x = c(1:2, NA), id = 1:3)
+  expect_equal(db |> slice_max(x, n = 3) |> pull(), 2:1)
+  expect_equal(db |> slice_min(x, n = 3) |> pull(), 1:2)
+})
+
 test_that("slice_* can use data masking pronouns", {
   lf <- lazy_frame(x = c(1, 1, 2), id = c(1, 2, 3))
   x <- -1L
@@ -47,7 +62,7 @@ test_that("slice_sample errors when expected", {
 
   # Can't see how to test this, but interactive experimentation
   # shows that it doesn't always return the same result
-  expect_error(db |> slice_sample() |> pull(), NA)
+  expect_no_error(db |> slice_sample() |> pull())
 
   expect_snapshot(error = TRUE, db |> slice_sample(replace = TRUE))
   expect_snapshot(error = TRUE, db |> slice_sample(weight_by = x))