+++ NEWS.md
@@ -1,5 +1,6 @@
 # dbplyr (development version)
 
+* `slice_*()` now handles missing values in line with the documentation, i.e. they are always removed (#1599).
 * Internal testing functions `src_test()`, `test_frame()` and `test_load()`, `test_register_src()` and `test_register_con()` have been removed.
 * A `filter()` followed by a `summarise()` is once again inlined correctly (#1707).
 * `src_memdb()` and `tbl_memdb()` are deprecated; use `memdb()` and `copy_to(memdb(), df)` instead. New `local_memdb_frame()` for use in tests.